Move grid into its own folder
