+++ src/grid/grid-cell.js
