+++ src/grid/index.js
@@ -1,0 +1,5 @@
+export { default as ColumnHeader } from './column-header.js';
+export { default as GridCell } from './grid-cell.js';
+export { default as Grid } from './grid.js';
+export { default as RowHeaders } from './row-headers.js';
+export { default as Row } from './row.js';
